renames value objects, exception messages, functions
